refactor: Move API code to the `server` directory
This just allows us to split the code easier, should also reduce the
amount of merge conflicts if someone was to edit index.ts in their PR.
