+++ src/server/routes/api/v1/spotify/auth.ts
@@ -1,5 +1,5 @@
 import { FastifyRequest } from "fastify"
-import { authorize } from "../../../../lib/spotify"
+import { authorize } from "../../../../../lib/spotify"
 
 interface AuthQuery {
     code: string
@@ -12,6 +12,6 @@ export async function spotifyAuthRoute(request: FastifyRequest<{ Body: AuthQuery
     if (result.ok) {
         return result.value
     } else {
-        return { "message": result.error.error_description }
+        return { message: result.error.error_description }
     }
 }
