+++ src/server/routes/api/v1/themes/list.ts
@@ -1,13 +1,7 @@
-import { FastifyRequest } from "fastify"
-import prisma from "../../../../util/prisma"
+import prisma from "../../../../../util/prisma"
 
-type Params = {
-    id: string
-}
-
-export async function themesIdRoute(request: FastifyRequest<{ Params: Params }>) {
-    const theme = await prisma.theme.findFirst({
-        where: { id: request.params.id },
+export async function themesListRoute() {
+    const themes = await prisma.theme.findMany({
         include: {
             author: {
                 select: {
@@ -27,9 +21,8 @@ export async function themesIdRoute(request: FastifyRequest<{ Params: Params }>)
         }
     })
 
-    if (!theme) return { ok: false, message: "Unknown theme!" }
     return {
         ok: true,
-        theme
+        themes: themes
     }
 }
